feat: optionally exclude non-protein coding genes from mapping dump  (#426)
Around 50% of the gene records we have are probably never going to be
used by varcat (lncRNA, etc) and may or may not end up polluting the
autocomplete results a lot -- so I stubbed this out while I was thinking
about it. We could be more hardcore about the specific types we select
but I think this is pretty close to what we'd ultimately want.